Consolidate the duplicate getRelativeTimeString function.
