+++ src/lib/components/feed/entry.svelte
@@ -5,6 +5,8 @@
 	import { goto } from '$app/navigation';
 	import { page } from '$app/stores';
 
+	import { getRelativeTimeString } from '$lib/date-time';
+
 	import CommentDialog from './comment-dialog.svelte';
 	import FileDialog from './file-dialog.svelte';
 
@@ -42,39 +44,6 @@
 				// Ignored, this could be due to the user cancelling the share activity
 			}
 		}
-	};
-
-	const getRelativeTimeString = (date: Date | number) => {
-		// Allow dates or times to be passed
-		const timeMs = typeof date === 'number' ? date : date.getTime();
-
-		// Get the amount of seconds between the given date and now
-		const deltaSeconds = Math.round((timeMs - Date.now()) / 1000);
-
-		// Array reprsenting one minute, hour, day, week, month, etc in seconds
-		const cutoffs = [60, 3600, 86400, 86400 * 7, 86400 * 30, 86400 * 365, Infinity];
-
-		// Array equivalent to the above but in the string representation of the units
-		const units: Intl.RelativeTimeFormatUnit[] = [
-			'second',
-			'minute',
-			'hour',
-			'day',
-			'week',
-			'month',
-			'year'
-		];
-
-		// Grab the ideal cutoff unit
-		const unitIndex = cutoffs.findIndex((cutoff) => cutoff > Math.abs(deltaSeconds));
-
-		// Get the divisor to divide from the seconds. E.g. if our unit is "day" our divisor
-		// is one day in seconds, so we can divide our seconds by this to get the # of days
-		const divisor = unitIndex ? cutoffs[unitIndex - 1] : 1;
-
-		// Intl.RelativeTimeFormat do its magic
-		const rtf = new Intl.RelativeTimeFormat('en', { numeric: 'auto' });
-		return rtf.format(Math.floor(deltaSeconds / divisor), units[unitIndex]);
 	};
 </script>
